Remove enum dependency on common.
This allows common to use enums.
