+++ tests/test_common.py
@@ -20,13 +20,11 @@ from pex.common import (
     is_exe,
     is_script,
     open_zip,
-    qualified_name,
     safe_mkdir,
     safe_open,
     temporary_dir,
     touch,
 )
-from pex.compatibility import PY2
 from pex.typing import TYPE_CHECKING
 
 try:
@@ -416,46 +414,6 @@ def test_is_script(tmpdir):
     assert is_script(exe, check_executable=False)
     assert not is_script(exe)
     assert not is_exe(exe)
-
-
-def test_qualified_name():
-    # type: () -> None
-
-    expected_str_type = "{module}.str".format(module="__builtin__" if PY2 else "builtins")
-    assert expected_str_type == qualified_name(str), "Expected builtin types to be handled."
-    assert expected_str_type == qualified_name(
-        "foo"
-    ), "Expected non-callable objects to be identified via their types."
-
-    assert "pex.common.qualified_name" == qualified_name(
-        qualified_name
-    ), "Expected functions to be handled"
-
-    assert "pex.common.AtomicDirectory" == qualified_name(
-        AtomicDirectory
-    ), "Expected custom types to be handled."
-    expected_prefix = "pex.common." if PY2 else "pex.common.AtomicDirectory."
-    assert expected_prefix + "finalize" == qualified_name(
-        AtomicDirectory.finalize
-    ), "Expected methods to be handled."
-    assert expected_prefix + "work_dir" == qualified_name(
-        AtomicDirectory.work_dir
-    ), "Expected @property to be handled."
-
-    expected_prefix = "pex.common." if PY2 else "pex.common.PermPreservingZipFile."
-    assert expected_prefix + "zip_entry_from_file" == qualified_name(
-        PermPreservingZipFile.zip_entry_from_file
-    ), "Expected @classmethod to be handled."
-
-    class Test(object):
-        @staticmethod
-        def static():
-            pass
-
-    expected_prefix = "test_common." if PY2 else "test_common.test_qualified_name.<locals>.Test."
-    assert expected_prefix + "static" == qualified_name(
-        Test.static
-    ), "Expected @staticmethod to be handled."
 
 
 def test_find_site_packages(tmpdir):
